fix the problems caused by solving conflicts.

+++ server/src/main/java/com/tripco/t04/planner/NearestNeighbor.java
@@ -41,7 +41,7 @@ public class NearestNeighbor extends Optimize {
     }
 
 
-    protected int iterate(int[] candidate, int minDist, int[][] latice) {
+    protected long iterate(int[] candidate, long minDist, long[][] latice) {
         return 0;
     }
 }
